+++ Text/Inflections/Humanize.hs
@@ -9,32 +9,34 @@
 --
 -- Conversion to “humanized” phrases.
 
-module Text.Inflections.Humanize (humanize) where
+{-# LANGUAGE CPP               #-}
+{-# LANGUAGE OverloadedStrings #-}
 
-import Text.Inflections.Parse.Types (Word(..))
+module Text.Inflections.Humanize
+  ( humanize )
+where
 
-import Data.Char (toUpper)
+import Data.Text (Text)
+import Text.Inflections.Parse.Types
+import qualified Data.Text as T
 
-import Prelude (String, Bool(..), (.), map, zip, ($), unwords, repeat)
+#if MIN_VERSION_base(4,8,0)
+import Prelude hiding (Word)
+#else
+import Control.Applicative
+#endif
 
--- |Capitalizes the first word and turns underscores into spaces. Like titleize,
--- this is meant for creating pretty output.
+-- |Capitalizes the first word and turns underscores into spaces. Like
+-- 'Text.Inflections.Titleize.titleize', this is meant for creating pretty
+-- output.
 --
 -- >>> humanize [ Word "foo", Acronym "bar", Word "bazz" ]
 -- "Foo bar bazz"
 humanize
   :: [Word] -- ^ List of Words, first of which will be capitalized
-  -> String -- ^ The humanized output
-humanize = unwords . map caseForWord . isFirstList
-
--- |Returns list with Bool indicating if an element is first.
-isFirstList :: [a] -> [(a, Bool)]
-isFirstList xs = zip xs $ True : repeat False
-
--- | Convert given 'Word' to capitalized 'String' when associated Boolean
--- value is 'True'.
-caseForWord :: (Word, Bool) -> String
-caseForWord (Word (c:cs), True)  = toUpper c : cs
-caseForWord (Word s, False)      = s
-caseForWord (Word [], _)         = []
-caseForWord (Acronym s, _)       = s  -- Acronyms are left intact
+  -> Text   -- ^ The humanized output
+humanize xs' =
+  case mapWord (T.replace "_" " ") <$> xs' of
+    []     -> ""
+    (x:xs) -> T.unwords $ T.toTitle x : (T.toLower <$> xs)
+{-# INLINE humanize #-}