+++ Text/Inflections/Parse/SnakeCase.hs
@@ -11,13 +11,16 @@
 
 {-# LANGUAGE CPP #-}
 
-module Text.Inflections.Parse.SnakeCase ( parseSnakeCase )
+module Text.Inflections.Parse.SnakeCase
+  ( parseSnakeCase )
 where
 
+import Data.Text (Text)
 import Text.Inflections.Parse.Acronym (acronym)
 import Text.Inflections.Parse.Types (Word(..))
 import Text.Megaparsec
-import Text.Megaparsec.String
+import Text.Megaparsec.Text
+import qualified Data.Text as T
 
 #if MIN_VERSION_base(4,8,0)
 import Prelude hiding (Word)
@@ -33,17 +36,20 @@ import Control.Applicative
 -- Left "(unknown)" (line 1, column 4):
 -- unexpected 'B'
 parseSnakeCase
-  :: [String]          -- ^ Collection of acronyms
-  -> String            -- ^ Input
+  :: [Text]            -- ^ Collection of acronyms
+  -> Text              -- ^ Input
   -> Either (ParseError Char Dec) [Word] -- ^ Result of parsing
 parseSnakeCase acronyms = parse (parser acronyms) ""
 
 parser
-  :: [String] -> Parser [Word]
+  :: [Text]
+  -> Parser [Word]
 parser acronyms = do
   ws <- (acronym acronyms <|> word) `sepBy` char '_'
   eof
   return ws
+{-# INLINE parser #-}
 
 word :: Parser Word
-word = Word <$> (some lowerChar <|> some digitChar)
+word = Word . T.pack <$> (some lowerChar <|> some digitChar)
+{-# INLINE word #-}